TimeoutCheck re-factor, NuGet updates
Applied null pattern to TimeoutCheck, wired everything up in AutoFac module, preliminary work in ConfirmationTimeoutChecker.

Updated NuGet modules (AutoFac and Microsoft Task etc...) across solution.

+++ Shield/Data/Models/IApplicationSettingsModel.cs
@@ -9,7 +9,7 @@ namespace Shield.Data.Models
         int CommandTypeSize { get; set; }
         char Separator { get; set; }
         char Filler { get; set; }
-        long ConfirmationTimeout { get; set; }
-        long CompletitionTimeout { get; set; }
+        int ConfirmationTimeout { get; set; }
+        int CompletitionTimeout { get; set; }
     }
 }
